requirements.txt
azure-functions
sqlalchemy==1.3.18
pyodbc==4.0.26